+++ src/browser/components/ProjectFiles/projectFilesConstants.ts
